+++ HISTORY.rst
@@ -5,9 +5,15 @@ History
 Unreleased
 ==========
 
+.. warning::
+
+   If your setup includes remote configured to be accessed over SSH, you may
+   need to reinstall ``libgit2`` with ``brew reinstall libgit2 --with-libssh2``.
+
 * Added facilities to integrate with the current shell, for stuff like cd'ing
   to other directories or activating virtual environments.
 * Added a ``--version`` option to ``lancet``.
+* Fetch latest changes from origin before creating new working branches (#1).
 
 0.3 - 2014-12-30
 ================